Move PRSC related files into prsc folder
